refactor: make an Icon component
