+++ src/App.vue
@@ -38,8 +38,8 @@ import { useStore } from "vuex";
 import BeatIndicator from "./components/beat/BeatIndicator.vue";
 import { useRegisterSW } from "virtual:pwa-register/vue";
 import Aside from "./components/Aside.vue";
-import Volume from "./components/settings/volume/Volume.vue";
-import Sound from "./components/settings/sound/Sound.vue";
+import Volume from "./components/settings/Volume.vue";
+import Sound from "./components/settings/Sound.vue";
 import FirstBeat from "./components/settings/FirstBeat.vue";
 import LargeIncrement from "./components/settings/LargeIncrement.vue";
 import BpmRange from "./components/settings/BpmRange.vue";
