+++ src/components/icons/IconClick.vue
