+++ src/components/icons/IconHiHat.vue
