+++ src/components/icons/IconKick.vue
