+++ src/components/icons/IconSticks.vue
