+++ src/components/icons/IconVolume.vue
