+++ src/components/icons/IconVolumeMute.vue
